% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_jobs.R
\name{search_jobs}
\alias{search_jobs}
\title{Query jobs}
\usage{
search_jobs(query, limit = 20)
}
\arguments{
\item{query}{A text query to search}

\item{limit}{Number of observations to return}
}
\description{
Query jobs
}
